pyproject.toml
[project]
name = "mini-diffusion-lm"
version = "0.1.0"
description = "A small discrete diffusion language model"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1",
    "transformers>=4.36",
    "datasets>=2.16",
    "pyyaml>=6.0",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = ["wandb", "tensorboard", "matplotlib"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
